Removed handleListQuery function and unused types
