+++ src/pages/Providers/index.vue
@@ -7,20 +7,12 @@
 </template>
 
 <script setup lang="ts">
-import { getListOfProviders } from '../../actions';
-import { handleListQuery } from '../../utils/apollo';
+import { handleProvidersList } from './utils';
 import ProvidersLayout from './Providers.layout.vue';
 import './Providers.scss';
 
-const {
-  items: providers,
-  query,
-  currentPage,
-  isLoading,
-  searchText,
-  order_by,
-  getItemsList: getProvidersList,
-} = handleListQuery(getListOfProviders);
+const { providers, getProvidersList, isLoading, query, searchText } =
+  handleProvidersList();
 
-await getProvidersList(currentPage.value, query.value, order_by.value);
+await getProvidersList(query.value);
 </script>
