+++ src/utils/apollo.types.ts
@@ -1,37 +1,3 @@
-// Functions types
-export interface actionCallbackReturnTypes<Return> {
-  items: Return[];
-  totalItems: number;
-}
-
-export interface orderByParamTypes {
-  label?: string;
-  content?: string;
-  column?: string;
-}
-
-export interface orderByGraphQLParamType {
-  [key: string]: string;
-}
-
-export interface actionCallbackParamsTypes {
-  limit?: number;
-  offset?: number;
-  query?: string;
-  order_by?: orderByGraphQLParamType;
-}
-
-export type actionCallbackType = <
-  ReturnType
->({}: actionCallbackParamsTypes) => Promise<
-  actionCallbackReturnTypes<ReturnType>
->;
-
-export type ActionCallbackType = {
-  items: object[];
-  totalItems: number;
-};
-
 // Database Models
 // Users
 export interface Users {
